+++ src/ic_registry/tests/inventory/mod.rs
@@ -1,0 +1,4 @@
+mod data_center;
+mod node;
+mod node_operator;
+mod node_provider;